+++ tests/python/test_zfs_repository.py
@@ -34,14 +34,23 @@ class TestZfsRepositoryReads(unittest.TestCase):
         repo._run = lambda *a, **k: result
         return repo
 
-    def test_list_pools_parses_six_columns(self):
-        repo = self._repo("tank\tONLINE\t10T\t5T\t5T\t75%\n")
+    def test_list_pools_parses_seven_columns(self):
+        repo = self._repo("tank\tONLINE\t10T\t5T\t5T\t75%\t-\n")
         pools = repo.list_pools()
         self.assertEqual(len(pools), 1)
-        self.assertEqual(pools[0], PoolRow("tank", "ONLINE", "10T", "5T", "5T", "75%"))
+        self.assertEqual(
+            pools[0],
+            PoolRow("tank", "ONLINE", "10T", "5T", "5T", "75%", "-"),
+        )
+
+    def test_list_pools_parses_active_checkpoint(self):
+        repo = self._repo("tank\tONLINE\t10T\t5T\t5T\t75%\t50G\n")
+        pools = repo.list_pools()
+        self.assertEqual(len(pools), 1)
+        self.assertEqual(pools[0].ckpoint, "50G")
 
     def test_list_pools_ignores_blank_lines(self):
-        repo = self._repo("\ntank\tONLINE\t10T\t5T\t5T\t75%\n\n")
+        repo = self._repo("\ntank\tONLINE\t10T\t5T\t5T\t75%\t-\n\n")
         self.assertEqual(len(repo.list_pools()), 1)
 
     def test_list_pools_full_parses_nine_columns(self):